rearranges scripts so Balsam Apps are defined in another script
