+++ libensemble/tests/scaling_tests/balsam_forces/run_libe_forces_balsam.py
@@ -6,6 +6,8 @@ from libensemble import Ensemble
 from libensemble.executors import BalsamExecutor
 from balsam.api import ApplicationDefinition
 
+BALSAM_SITE = "jln_theta"
+
 THIS_SCRIPT_ON_THETA = any(
     [i in socket.gethostname() for i in ["theta", "nid0"]]
 )  # Is this running on a personal machine, or a compute node?
@@ -30,7 +32,7 @@ forces.sim_specs["user"].update(
 
 forces.persis_info.add_random_streams()
 
-apps = ApplicationDefinition.load_by_site("jln_theta")
+apps = ApplicationDefinition.load_by_site(BALSAM_SITE)
 RemoteForces = apps["RemoteForces"]
 
 exctr = BalsamExecutor()
